+++ src/PostgREST/Types.hs
@@ -31,12 +31,17 @@ import Protolude.Conv       (toS)
 -- | Enumeration of currently supported response content types
 data ContentType = CTApplicationJSON | CTSingularJSON
                  | CTTextCSV | CTTextPlain
-                 | CTOpenAPI | CTOctetStream
+                 | CTOpenAPI | CTUrlEncoded | CTOctetStream
                  | CTAny | CTOther ByteString deriving (Show, Eq)
 
 -- | Convert from ContentType to a full HTTP Header
 toHeader :: ContentType -> Header
-toHeader ct = (hContentType, toMime ct <> "; charset=utf-8")
+toHeader ct = (hContentType, toMime ct <> charset)
+  where
+    charset = case ct of
+      CTOctetStream -> mempty
+      CTOther _     -> mempty
+      _             -> "; charset=utf-8"
 
 -- | Convert from ContentType to a ByteString representing the mime type
 toMime :: ContentType -> ByteString
@@ -45,6 +50,7 @@ toMime CTTextCSV         = "text/csv"
 toMime CTTextPlain       = "text/plain"
 toMime CTOpenAPI         = "application/openapi+json"
 toMime CTSingularJSON    = "application/vnd.pgrst.object+json"
+toMime CTUrlEncoded      = "application/x-www-form-urlencoded"
 toMime CTOctetStream     = "application/octet-stream"
 toMime CTAny             = "*/*"
 toMime (CTOther ct)      = ct
@@ -58,6 +64,7 @@ decodeContentType ct = case BS.takeWhile (/= BS.c2w ';') ct of
   "application/openapi+json"          -> CTOpenAPI
   "application/vnd.pgrst.object+json" -> CTSingularJSON
   "application/vnd.pgrst.object"      -> CTSingularJSON
+  "application/x-www-form-urlencoded" -> CTUrlEncoded
   "application/octet-stream"          -> CTOctetStream
   "*/*"                               -> CTAny
   ct'                                 -> CTOther ct'