+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - #1094, Fix expired JWTs starting an empty transaction on the db - @steve-chavez
  - #1162, Fix location header for POST request with select= without PK - @wolfgangwalther
  - #1585, Fix error messages on connection failure for localized postgres on Windows - @wolfgangwalther
+ - #1636, Fix `application/octet-stream` appending `charset=utf-8` - @steve-chavez
 
 ### Changed